+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 CHANGELOG
 =========
 
+* bugfix #351 fix handling of boolean in import/export.
+
 1.3.5
 -----